+++ YouboraLib/plugin/YBPlugin.m
@@ -1729,9 +1729,9 @@
     return self.options.appReleaseVersion;
 }
 
-- (NSString *) getFingerprint {
+- (NSString *) getDeviceUUID {
     #if TARGET_OS_IPHONE==1
-        if (UIDevice.currentDevice.identifierForVendor) {
+        if (UIDevice.currentDevice.identifierForVendor && !self.options.deviceIsAnonymous) {
             return UIDevice.currentDevice.identifierForVendor.UUIDString;
         }
     #endif
@@ -2497,8 +2497,13 @@
 
 - (void) eventListenerDidReceivetoBack: (NSNotification*)uselessNotification {
     if(self.options.autoDetectBackground){
-        if (self.adsAdapter != nil && (self.adsAdapter.flags.started || self.adsAdapter.flags.adInitiated)) {
-            [self.adsAdapter fireStop];
+        if (self.adsAdapter != nil) {
+            if (self.adsAdapter.flags.started || self.adsAdapter.flags.adInitiated) {
+                [self.adsAdapter fireStop];
+            }
+            if (self.adsAdapter.flags.adBreakStarted) {
+                [self.adsAdapter fireAdBreakStop];
+            }
         }
         
         if (self.isInitiated && (self.adapter == nil || !self.adapter.flags.started)) {