+++ CHANGELOG.md
@@ -1,4 +1,7 @@
+
 ## [6.5.1] - 2019-
+###Added
+- Break stop request when going to background
 ###Replaced
 - Now deviceUUID is used instead of Fingerprint